Move test helpers into cucumber helper modules Remove old test api files and requires
